+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release history
 
+## trio-util (pending)
+### Added
+- `@trio_async_generator` is a decorator which adapts a generator containing
+  Trio constructs for safe use.  (Normally, it's not allowed to yield from a
+  nursery or cancel scope when implementing async generators.)
+
 ## trio-util 0.2.0 (2020-09-09)
 ### Added
 - `AsyncValue.wait_value() / wait_transition()` additionally accept a plain